[experimental] Mock transport instead of the underlying `fetch` primitive (#1222)
## Summary

From the perspective of the RPC transport, it's not interesting that the wire transport uses `fetch`. All that's interesting is that the wire transport returns responses.

In this PR, we mock at the wire transport level, not at the underlying `fetch` API level. The `fetch` API level is already well tested by `http-transport-test.ts`

## Test Plan

```
pnpm test:unit:browser
pnpm test:unit:node
```