refactor(experimental): jest setup/teardown functions for test validator (#1228)
## Summary
Any test that relies on a test validator running on localhost can use these Jest setup/teardown functions. They take care of starting the validator, waiting for it to come alive, and tearing it down at the end of the test.

## Test Plan

See the first PR that makes use of it.