refactor(experimental): make `abortSignal` optional on send-and-confirm-transaction helpers (#1763)
# Summary

It's conceivable that, despite the fact that the send-and-confirm helpers use subscriptions internally, you might want to be lazy and not manage cancellation. This is so because transaction confirmation is short-lived-ish.

To save you a few keystrokes, and at your own peril, we make `abortSignal` optional in the method signature of `sendAndConfirmTransaction` and `sendAndConfirmDurableNonceTransaction`.

# Test Plan

```
cd packages/library/
pnpm test:unit:browser
pnpm test:unit:node
```