refactor(experimental): Use the native `fetch` built into Node 18+ (#1802)
# Summary

OK, so here's what I'm thinking.

`node-fetch` is in this weird state where:

- they refuse to build CommonJS bundles for the 3.x line
- the 2.x line supports `HttpAgent`
  - note that browsers will never support `HttpAgent`
- the 2.x line does not support HTTP/2

This means that we're taking on a huge dependency and exposing its config through web3.js, but will never end up in the ideal state (HTTP/2 support everywhere).

In this PR, we throw in the towel and make the _default_ transport that ships with web3.js API compatible with the `fetch` in both Node and browsers.

From that point:

- We could write our own HTTP/2 transport that's API compatible with `fetch()` and ship that
- Someone _else_ could write an all-singing, all-dancing transport _specifically_ for Node and open source it

Addresses #1680.