refactor(experimental): a sham for `Transaction` (#1890)
# Summary

This is a sham that lets you wrap some transaction bytes (in wire format) in a `Transaction`-like class that will vend you the signature. This is the minimum one might like to do given a sent transaction that they just want the signature of (eg. for lookup on a block explorer)

# Test Plan

```
cd packages/library-legacy-sham
pnpm test:unit:browser
pnpm test:unit:node
```