refactor(experimental): codecs: refactor `message` out of codec assertions (#2200)
As mentioned in
[this PR comment](https://github.com/solana-labs/solana-web3.js/pull/2188#discussion_r1501789028),
if you need to provide some kind of "reason" or "message" to an error, it's
usually a smell.

Now that we have custom error codes and can perform proper error handling, we
can replace this mechanism for piping error messages (effectively "reasons") to
fixed codec assertions!

This change refactors out `message` from `assertIsFixed`. In a later commit,
I'll replace the thrown error with a coded `SolanaError` and add the necessary
`catch` statements downstream.