Publish the 2.0.0 line with `changesets` now (#2274)
# Summary

This PR aims to switch the experimental library over to Changesets for publishing.

Every commit to the experimental library will now create/update a changesets PR. When we want to release, we land that PR and Changesets does the following:

1. Increments the versions of everything in unison to `2.0.0-preview.{X+1}`
2. Commits the version numbers and the `CHANGELOG.md` files to the repo
3. Sets `hasChangesets` to `true` in the CI/CD GitHub Actions workflow, which causes our `publish-packages` script to publish the new versions.

# Test Plan

Have I tested this? Nope. The most I've done is to do this locally:

1. `pnpm changeset` – This is what the GitHub bot encourages you to do on PRs
2. `pnpm changeset version` – This is essentially what the Action in `cicd.yml` does

That looked good to me, but whether it will all work in production is anyone's guess.